change from int to double

+++ shisoku.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 
 int main(){
-	int a, b;
+	double a, b;
 
 	a=2;
 	b=3;
 
-	printf("a+b=%d\n",a+b);
+	printf("a+b=%f\n",a+b);
 
-	printf("a-b=%d\n", a-b);
+	printf("a-b=%f\n", a-b);
 
-	printf("a*b=%d\n",a*b);
+	printf("a*b=%f\n",a*b);
 
-	printf("a/b=%d\n", a/b);
+	printf("a/b=%f\n", a/b);
 
 	return 0;
 }
